refactor(strategy): split dataclasses into separate files and add name field
- Extract StrategyCondition to strategy_condition.py
- Extract StrategyConfig to strategy_config.py
- Add name parameter to StrategyManager, StrategyExecutor, StrategyFetcher
- Update __init__.py exports
- Standardized logger naming to f"{self.__class__.__name__}_{self.name}"
