+++ src/strategy/strategy_condition.py
@@ -1,0 +1,9 @@
+# STANDARD LIBRARY
+from dataclasses import dataclass
+from typing import Any, Dict
+
+
+@dataclass
+class StrategyCondition:
+    type: str
+    payload: Dict[str, Any]
